+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Tantivy 0.14.0
 - Migrated tantivy error from the now deprecated `failure` crate to `thiserror` #760. (@hirevo)
 - API Change. Accessing the typed value off a `Schema::Value` now returns an Option instead of panicking if the type does not match. 
 
+Tantivy 0.13.2
+===================
+Bugfix. Acquiring a facet reader on a segment that does not contain any 
+doc with this facet returns `None`. (#896)
+
 Tantivy 0.13.1
 ===================
 Made `Query` and `Collector` `Send + Sync`.